Write a JSON representation of the schema to the schema jar
This will allow us to switch to parsing the JSON from a dependency rather than re-parsing the dependency's schema. Why? We want to switch to AST-parsing the schema which is something we can only do on sources. Thus, anything we want from a
dependency needs to come in another way.